pyproject.toml
[project]
name = "django-s3-session-storage"
version = "0.1.0"
description = "A Django S3 storage backend that supports dependency injection and shared boto3.Session reuse."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    # TODO : Update with your actual name and email
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["django", "storages", "aws", "s3", "boto3", "session", "dependency-injection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "boto3>=1.26.0",
    "django-storages>=1.14.0",
    "django>=3.2",
]

[project.urls]
# TODO : Update these URLs with your actual project links
Homepage = "https://github.com/yourname/django-s3-session-storage"
Repository = "https://github.com/yourname/django-s3-session-storage"
Documentation = "https://github.com/yourname/django-s3-session-storage#readme"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

